Drawer as different component Hamburger With Search as different component SignOutResponse change common deleted(unnecessary file)
