+++ src/global/graphqls/SignOut.ts
@@ -1,9 +1,7 @@
 import { gql } from 'apollo-boost';
 
 export interface SignOutResponse {
-  loginUser: {
-    accessToken: string;
-  };
+  logout: boolean;
 }
 
 export const SIGN_OUT = gql`
